requirements.txt
django>=1.9.6
django-model-utils>=2.0
# Additional requirements go here
facebook-sdk
